[Edit] CustomBottomSheet 리팩토링 #85
