+++ FinalHrmiProjects/Resource/PresentationDetent +.swift
@@ -1,0 +1,14 @@
+//
+//  PresentationDetent +.swift
+//  FinalHrmiProjects
+//
+//  Created by phang on 2/3/24.
+//
+
+import SwiftUI
+
+// sheet 높이 직접 지정 
+extension PresentationDetent {
+    static let drinkInfo = Self.height(346)
+    static let displaySetting = Self.height(244)
+}
