+++ FinalHrmiProjects/View/Mypage/SettingView.swift
@@ -103,13 +103,15 @@ struct SettingView: View {
 				CustomDivider()
 				Spacer()
 			}
-			
 			// 화면 모드 설정 클릭 시 띄워지는 CustomBottomSheet
 			// BottomSheetType -> .displaySetting
-			CustomBottomSheet(isShowingSheet: $isShowingSheet) {
-				BottomSheetContentView(optionNameList: optionNameList, isShowingSheet: $isShowingSheet, selectedSortingOption: $selectedSortingOption, text: BottomSheetType.displaySetting.description)
-			}
-			
+            .sheet(isPresented: $isShowingSheet) {
+                CustomBottomSheetContent(optionNameList: optionNameList, isShowingSheet: $isShowingSheet, selectedSortingOption: $selectedSortingOption, bottomSheetType: .displaySetting)
+                    .presentationDetents([.displaySetting])
+                    .presentationDragIndicator(.hidden) // 시트 상단 인디케이터 비활성화
+                    .interactiveDismissDisabled() // 내려서 닫기 비활성화
+            }
+    
 			// 로그아웃 버튼 클릭 시 띄워지는 CustomAlert
 			if isLogoutClicked {
 				CustomSettingDialog(message: "로그아웃 하시겠습니까?", leftButtonLabel: "취소", leftButtonAction: {
